msvcrt: Use iob_func instead of p__iob on non-x86.

+++ dlls/msvcrt/main.c
@@ -67,10 +67,10 @@ BOOL WINAPI DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved)
 {
   thread_data_t *tls;
 
-  TRACE("(%p, %s, %p) pid(%x), tid(%x), tls(%ld)\n",
+  TRACE("(%p, %s, %p) pid(%x), tid(%x), tls(%u)\n",
         hinstDLL, msvcrt_get_reason(fdwReason), lpvReserved,
         GetCurrentProcessId(), GetCurrentThreadId(),
-        (long)msvcrt_tls_index);
+        msvcrt_tls_index);
 
   switch (fdwReason)
   {
